workspace/python/requirements.txt
# ============================================
# NexusForge Platform - Python Dependencies
# ============================================
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.2.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
tenacity==8.2.3
pytz==2023.3.post1
